+++ external/Box2D/Collision/b2CollideCircle.cpp
@@ -31,9 +31,9 @@ void b2CollideCircles(
 	b2Vec2 pB = b2Mul(xfB, circleB->m_p);
 
 	b2Vec2 d = pB - pA;
-	float32 distSqr = b2Dot(d, d);
-	float32 rA = circleA->m_radius, rB = circleB->m_radius;
-	float32 radius = rA + rB;
+	qreal distSqr = b2Dot(d, d);
+	qreal rA = circleA->m_radius, rB = circleB->m_radius;
+	qreal radius = rA + rB;
 	if (distSqr > radius * radius)
 	{
 		return;
@@ -61,15 +61,15 @@ void b2CollidePolygonAndCircle(
 
 	// Find the min separating edge.
 	int32 normalIndex = 0;
-	float32 separation = -b2_maxFloat;
-	float32 radius = polygonA->m_radius + circleB->m_radius;
+	qreal separation = -b2_maxFloat;
+	qreal radius = polygonA->m_radius + circleB->m_radius;
 	int32 vertexCount = polygonA->m_vertexCount;
 	const b2Vec2* vertices = polygonA->m_vertices;
 	const b2Vec2* normals = polygonA->m_normals;
 
 	for (int32 i = 0; i < vertexCount; ++i)
 	{
-		float32 s = b2Dot(normals[i], cLocal - vertices[i]);
+		qreal s = b2Dot(normals[i], cLocal - vertices[i]);
 
 		if (s > radius)
 		{
@@ -103,8 +103,8 @@ void b2CollidePolygonAndCircle(
 	}
 
 	// Compute barycentric coordinates
-	float32 u1 = b2Dot(cLocal - v1, v2 - v1);
-	float32 u2 = b2Dot(cLocal - v2, v1 - v2);
+	qreal u1 = b2Dot(cLocal - v1, v2 - v1);
+	qreal u2 = b2Dot(cLocal - v2, v1 - v2);
 	if (u1 <= 0.0f)
 	{
 		if (b2DistanceSquared(cLocal, v1) > radius * radius)
@@ -138,7 +138,7 @@ void b2CollidePolygonAndCircle(
 	else
 	{
 		b2Vec2 faceCenter = 0.5f * (v1 + v2);
-		float32 separation = b2Dot(cLocal - faceCenter, normals[vertIndex1]);
+		qreal separation = b2Dot(cLocal - faceCenter, normals[vertIndex1]);
 		if (separation > radius)
 		{
 			return;
